+++ config/README.md
@@ -1,3 +1,5 @@
+TODO: UPDATE THIS
+
 # hubspot/local-dev-lib
 
 ## Config utils
