+++ config/__tests__/environment.test.ts
@@ -1,3 +1,5 @@
+// @TODO: update tests
+
 import { loadConfigFromEnvironment } from '../environment';
 import { ENVIRONMENT_VARIABLES } from '../../constants/environments';
 import { PERSONAL_ACCESS_KEY_AUTH_METHOD } from '../../constants/auth';
