+++ types/Config.ts
@@ -24,28 +24,7 @@ export type DeprecatedHubSpotConfigFields = {
   defaultMode?: CmsPublishMode;
 };
 
-// export interface CLIConfig_DEPRECATED {
-//   portals: Array<CLIAccount_DEPRECATED>;
-//   allowUsageTracking?: boolean;
-//   defaultPortal?: string | number;
-//   defaultMode?: CmsPublishMode; // Deprecated - left in to handle existing configs with this field
-//   defaultCmsPublishMode?: CmsPublishMode;
-//   httpTimeout?: number;
-//   env?: Environment;
-//   httpUseLocalhost?: boolean;
-// }
-
 export type Environment = ValueOf<typeof ENVIRONMENTS> | '';
-
-export type EnvironmentConfigVariables = {
-  apiKey?: string;
-  clientId?: string;
-  clientSecret?: string;
-  personalAccessKey?: string;
-  accountId?: number;
-  refreshToken?: string;
-  env?: Environment;
-};
 
 export type GitInclusionResult = {
   inGit: boolean;
